Work save

+++ Models/Spectrum.cs
@@ -294,7 +294,7 @@ namespace DiplomaMB.Models
                 var values = line.Split(',');
 
                 Wavelengths.Add(double.Parse(values[0], CultureInfo.InvariantCulture));
-                DataValues.Add(Convert.ToUInt32(values[1]));
+                DataValues.Add(Convert.ToDouble(values[1]));
             }
             Id = id;
             Name = Path.GetFileNameWithoutExtension(file_path);
